+++ scripts/generate-component-config/build-component-config.js
@@ -2,6 +2,7 @@ const path = require('path');
 const fs = require('fs');
 const kebabCase = require('lodash.kebabcase');
 const startCase = require('lodash.startcase');
+const { terraNameSpacedRepos, blackListedDirectories } = require('./build-configs');
 
 const componentImportNames = {};
 const packageConfigs = [];
@@ -51,7 +52,7 @@ const getImportName = (packageName, fileName, fileType) => {
 
 /** Builds the nested directory configuration for a found file.
   */
-const buildNestedComponentConfig = (nestedDirectories, fileConfig, fileType) => {
+const buildNestedComponentConfig = (nestedDirectories, fileConfig, fileType, packageName) => {
   let exampleConfig = fileConfig;
 
   // Reverse the order to build out the nested configuration from n to 0 directories deep
@@ -59,12 +60,16 @@ const buildNestedComponentConfig = (nestedDirectories, fileConfig, fileType) => 
 
   // Create the configuration for each directory
   nestedDirectories.forEach((dir) => {
-    const currentConfig = {
-      name: `'${startCase(dir)}'`,
-      path: `'/${kebabCase(dir)}'`,
-    };
-    currentConfig[`${fileType}`] = [exampleConfig];
-    exampleConfig = currentConfig;
+    if (!blackListedDirectories.includes(dir)) {
+      // If the packageName is a terra repository name, remove the repo prefix from the name and path
+      const directoryName = terraNameSpacedRepos.includes(packageName) ? dir.replace('terra-', '') : dir.replace(packageName, '');
+      const currentConfig = {
+        name: `'${startCase(directoryName)}'`,
+        path: `'/${kebabCase(directoryName)}'`,
+      };
+      currentConfig[`${fileType}`] = [exampleConfig];
+      exampleConfig = currentConfig;
+    }
   });
 
   return exampleConfig;
@@ -82,7 +87,13 @@ const buildComponentConfig = (foundFiles, repositoryName, outputPathDepth) => {
 
     // Get the example's package name
     let packageName = JSON.parse(fs.readFileSync(path.resolve(process.cwd(), 'package.json'))).name;
-    if (directory.includes('packages')) {
+    let monoRepoInstalledAsPackage = false;
+    if (directory.includes('node_modules') && directory.includes('packages')) {
+      // If a mono-repo is pulled in as a package, use it as the package name
+      // Note: spliting on seperator results in the first array element to be ''
+      packageName = directory.split('node_modules')[1].split(path.sep)[1];
+      monoRepoInstalledAsPackage = true;
+    } else if (directory.includes('packages')) {
       // The package name is the first directory name after packages.
       // Note: spliting on seperator results in the first array element to be ''
       packageName = directory.split('packages')[1].split(path.sep)[1];
@@ -119,13 +130,13 @@ const buildComponentConfig = (foundFiles, repositoryName, outputPathDepth) => {
 
     // Determine if the example needs nested configuration built
     const packageDirectories = directory.split(packageName)[1].split(path.sep);
-    const sliceAt = fileType === 'tests' || packageName.includes('-site') ? 3 : 2;
+    const sliceAt = (fileType === 'tests' && !monoRepoInstalledAsPackage) || packageName.includes('-site') ? 3 : 2;
     const nestedDirectories = packageDirectories.splice(sliceAt, packageDirectories.length);
 
     // Create the example's full configuration
     let exampleConfig = fileConfig;
     if (nestedDirectories.length) {
-      exampleConfig = buildNestedComponentConfig(nestedDirectories, exampleConfig, fileType);
+      exampleConfig = buildNestedComponentConfig(nestedDirectories, exampleConfig, fileType, packageName);
     }
 
     // Add the package configuration