+++ src/config/navigation.config.js
@@ -1,4 +1,4 @@
-module.exports = {
+const navConfig = {
   rootPath: '/site',
   navigation: {
     index: '/site/home',
@@ -6,7 +6,6 @@ module.exports = {
       path: '/site/home',
       text: 'Home',
       exampleType: 'home',
-      isStatic: true,
     }, {
       path: '/site/components',
       text: 'Components',
@@ -16,6 +15,9 @@ module.exports = {
       path: '/tests',
       text: 'Tests',
       exampleType: 'tests',
+      hasSubNav: true,
     }],
   },
 };
+
+export default navConfig;